+++ companions/thankyoujeffrey.zsh
@@ -1,0 +1,5 @@
+#!/usr/bin/env zsh
+
+emulate -L zsh
+script_dir=${0:A:h}
+exec "$script_dir/thankyoujeffrey.sh" "$@"